further reorganize code.
